feat(transport): add transport module with base class and jsonrpc and xmlrpc protocols
This commit adds a new module `transport` with a base class `BaseTransport`
and two subclasses `JsonRpcProtocol` and `XmlRpcProtocol`. The `BaseTransport`
class is an abstract base class that defines the interface for transport
protocols. The `JsonRpcProtocol` and `XmlRpcProtocol` classes implement the
transport protocols for JSON-RPC and XML-RPC respectively.

The `BaseTransport` class defines an abstract class method `from_settings`
that is used to create an instance of the transport protocol from an instance
of the `OdooSettings` class.

The `JsonRpcProtocol` and `XmlRpcProtocol` classes implement the `from_settings`
method to create an instance of the respective transport protocol from an
instance of the `OdooSettings` class.