feat(domain): add BaseDomain and CommonDomain classes to support common domain operations
The BaseDomain class is an abstract class that defines the basic structure of a domain class. It has an abstract method `_execute` that is used to execute a method on the domain and an abstract property `domain_name` that returns the name of the domain.

The CommonDomain class is a concrete implementation of the BaseDomain class that supports common domain operations. It has methods `version` and `authenticate` that execute the corresponding methods on the common domain.